refactor(profile): gate like/follow at comment parity, not per-proposal visibility
Rework of the target guard: instead of a bespoke draft/hidden/flagged
visibility check (which made likes stricter than comments and required
extracting getProposal's gate), gate a proposal like/follow the same way
commenting is — SUBMIT_PROPOSALS on the parent decision. Consistent with the
established engagement model, and getProposal is left untouched.

- assertProposalEngagementAccess: resolves proposal -> parent decision,
  asserts SUBMIT_PROPOSALS; no-op for public org/person targets
- revert getProposal extraction; drop assertProposalProfileVisible
- test rewritten: submit-capable caller may like, unrelated confirmed account
  may not
